refactor(conversation): tighten layout type ownership
